Minor optimizations have been made.
 Changes to be committed:
	modified:   ../lib/GA_parameterSearch/divideUsageVector.m
	modified:   ../lib/GA_parameterSearch/fitnessFunction.m


Former-commit-id: ed97cfff7cdb1f74a7f1ef9adf060c9690a5a8e1

+++ lib/GA_parameterSearch/divideUsageVector.m
@@ -41,5 +41,5 @@ function partIndex = divideUsageVector(usageVectorSize, DAY_PIECE)
 	partIndex(end,2) = partIndex(end,2) + md;
 	
 	% Transpose <partIndex>
-	partIndex = transpose(partIndex);
+	partIndex = uint16(transpose(partIndex));
 end
